pyproject.toml
[tool.poetry]
name = "sprinkler"
version = "0.1.0"
description = "Raspberry Pi Sprinkler ]["
authors = ["Jon Morgan <morgan.jon@gmail.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
django-environ = "^0.8.1"
Jinja2 = "^3.0.3"
djangorestframework = "^3.13.1"
requests = "^2.26.0"
loguru = "^0.5.3"
python-crontab = "^2.6.0"
"RPi.GPIO" = {version = "^0.7.1", markers = "sys_platform == 'linux'"}
gpiozero = "^1.6.2"
gunicorn = "^20.1.0"
click = "^8.1.3"
rich = "^12.4.0"
pylint-django = "^2.5.3"
yamllint = "^1.27.1"
flake8 = "^5.0.4"
black = "^22.6.0"
pycodestyle = "^2.9.1"
pylint = "^2.14.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-sugar = "^0.9.4"
coverage = "^6.2"
selenium = "^4.1.0"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
